requirements.txt
requests~=2.32.5
PyYAML~=6.0.3
fastapi~=0.128.0
pydantic~=2.12.5
Flask~=3.1.2
colorlog~=6.10.1